requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
openpyxl
python-dotenv
requests
langchain
langchain-google-genai
pydantic
